session: the handle's shape is the runtime's, and what is behind it is the backend's
`runtime/compute_session/mod.rs` held both halves of the compute segment: the
sequencing — when a session opens, what latches a `SequencingBlock`, when it
commits — and `ComputeSession` itself, whose `SessionRail` enum names
`metal::MetalSession` and whose three encode methods match on it.

The enum is right. `Backend` is not object-safe (its methods are generic over
`HostMemory + HostOps`), which is exactly why `SelectedBackend` is an enum too,
and `SessionRail`'s `Infallible` variant is what keeps the type one shape on a
build with no session-capable rail. What was wrong was the address: a neutral
runtime module was the one place in this device outside `backend/` that had to
name a rail to compile.

So `ComputeSession` and `SessionRail` move to `backend/compute_session.rs`,
beside the other type whose shape is neutral and whose contents are a rail's.
`runtime/compute_session` keeps `ComputeSegment`, `ensure_session`,
`apply_sequencing` and `finish_session` — it sequences a handle and never learns
which rail is behind it — and now names a rail only in its own `cfg(test)`
module, which `AGENTS.md` allows and which those tests need, since they open a
Metal encoder and assert what it does.

`runtime/compute_session/metal.rs` stays put: it is a rail file under the
runtime like `draw/metal` and `compute_exec/metal`, and moves with that unit,
not with this cut.

Verified: feature matrix all arms (counts unchanged), 2767 passed / 0 failed,
clippy unchanged on all three arms.
